fix(contract): fix possible staking overflow (#1510)
This commit introduces a test to check for overflow scenarios in the
stake function of the RewardsDistributionV2 contract. It simulates an
overflow condition using stdError and ensures that the correct revert
behavior is triggered.

Change the `totalStaked` and `stakedByDepositor` data types from
`uint256` to `uint96` across multiple contracts to optimize storage.
Ensure that the test suite properly bounds values within the new
`uint96` range.

+++ contracts/src/base/registry/facets/distribution/v2/RewardsDistribution.sol
@@ -377,7 +377,7 @@ contract RewardsDistribution is
   /// @inheritdoc IRewardsDistribution
   function stakedByDepositor(
     address depositor
-  ) external view returns (uint256 amount) {
+  ) external view returns (uint96 amount) {
     RewardsDistributionStorage.Layout storage ds = RewardsDistributionStorage
       .layout();
     amount = ds.staking.stakedByDepositor[depositor];
